% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2zotero.r
\name{r2zotero}
\alias{r2zotero}
\title{r -> Zotero}
\usage{
r2zotero(pkg = "base", manual_as_document = FALSE, verbose = TRUE)
}
\arguments{
\item{pkg}{Package to cite.}

\item{manual_as_document}{Whether to export \code{MANUAL} bibtex entries as
a \code{document} rather than \code{software}. Defaults to \code{FALSE}.}

\item{verbose}{Whether to print verbose output to console. Defaults to
\code{TRUE}.}
}
\value{
No return value. Function is called for its side effects.
}
\description{
Send citations from the \code{\link{citation}} function directly to Zotero
library through the Zotero JSON web API.
}
\section{Side effects}{

Imports the citations generated by \code{citation} to a local Zotero library.
}

\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
## send package references to Zotero
r2zotero("foreign")
r2zotero("lattice")
\dontshow{\}) # examplesIf}
}
\author{
Ethan Bass
}
